requirements.txt
pillow
Send2Trash